move state obs
