+++ pgdrive/obs/__init__.py
@@ -1,4 +1,5 @@
 from .image_obs import ImageObservation, ImageStateObservation
-from .observation_type import ObservationType, StateObservation, LidarStateObservation
-from .top_down_obs_multi_channel import TopDownMultiChannel
+from .observation_type import ObservationType
+from .state_obs import StateObservation, LidarStateObservation
 from .top_down_obs import TopDownObservation
+from .top_down_obs_multi_channel import TopDownMultiChannel
